+++ pkg/virtual/apiexport/schemas/builtin/builtin_test.go
@@ -1,0 +1,30 @@
+/*
+Copyright 2022 The KCP Authors.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package builtin
+
+import (
+	"testing"
+
+	"github.com/stretchr/testify/require"
+)
+
+func TestInit(t *testing.T) {
+	// NOTE(hasheddan): the length of APIResourceSchemas should be two less than
+	// the list of internal APIs due to the fact that v1beta1 and v1beta2 API
+	// versions are registered for FlowSchemas and PriorityLevelConfigurations.
+	require.Equal(t, len(BuiltInAPIs)-2, len(builtInAPIResourceSchemas))
+}
